+++ src/devlegate/__init__.py
@@ -4,6 +4,6 @@
 
 """Devlegate package."""
 
-__version__ = "0.5.4"
+__version__ = "0.5.5.dev0"
 
 __all__ = ["__version__"]